requirements.txt
pytubefix
llama-index-vector-stores-lancedb
llama-index-multi-modal-llms-openai
llama-index-embeddings-clip
llama_index
ftfy
regex
tqdm
-U openai-whisper
git+https://github.com/openai/CLIP.git
torch
torchvision
matplotlib
scikit-image
lancedb
moviepy
pytube
pydub
SpeechRecognition
ffmpeg-python
soundfile
streamlit